Tests - Use packages.3liz.org for fetching QGIS server plugins

+++ tests/add_server_plugins.sh
@@ -1,28 +1,31 @@
 #!/usr/bin/env bash
 
-echo "Remove legacy folders about qgis-plugin-manager"
-# For a few weeks
-rm -rf /srv/plugins/sources.list
-rm -rf /srv/plugins/.cache_qgis_plugin_manager
+# Remove legacy folders about qgis-plugin-manager
+if [ -d /srv/plugins/.cache_qgis_plugin_manager ]; then
+  rm -rf /srv/plugins/.cache_qgis_plugin_manager
+fi
+if [ -f /srv/plugins/sources.list ]; then
+  rm -f /srv/plugins/sources.list
+fi
 
-echo "QGIS Server Lizmap plugin"
+echo "QGIS Server Lizmap, WfsOutputExtension and AtlasPrint plugins"
+echo "Unstable from https://packages.3liz.org"
+# qgis-plugin-manager init
+echo "https://packages.3liz.org/pub/server-plugins-repository/unstable/plugins.[VERSION].xml" > /tmp/sources-plugin-manager.list
+qgis-plugin-manager update
+qgis-plugin-manager install "Lizmap server"
+qgis-plugin-manager install atlasprint
+qgis-plugin-manager install wfsOutputExtension
 
+# echo "QGIS Server Lizmap plugin"
 # Latest commit
-echo "Latest commit from https://packages.3liz.org"
-wget https://packages.3liz.org/pub/lizmap-server-qgis-plugin/lizmap_server.master.zip -O /tmp/lizmap_server.master.zip
+# echo "Latest commit from https://packages.3liz.org"
+# wget https://packages.3liz.org/pub/lizmap-server-qgis-plugin/lizmap_server.master.zip -O /tmp/lizmap_server.master.zip
 
 # Latest release
 # VERSION=1.1.1
 # echo "Stable release from GitHub"
 # wget https://github.com/3liz/qgis-lizmap-server-plugin/releases/latest/download/lizmap_server.${VERSION}.zip -O /tmp/lizmap_server.master.zip
 
-unzip -o /tmp/lizmap_server.master.zip -d /srv/plugins/
-rm /tmp/lizmap_server.master.zip
-
-echo "QGIS Server WfsOutputExtension and AlasPrint plugins"
-echo "Stable releases from http://plugins.qgis.org"
-qgis-plugin-manager init
-qgis-plugin-manager update
-# qgis-plugin-manager install "Lizmap server"
-qgis-plugin-manager install atlasprint
-qgis-plugin-manager install wfsOutputExtension
+# unzip -o /tmp/lizmap_server.master.zip -d /srv/plugins/
+# rm /tmp/lizmap_server.master.zip
